[Feat/#336] 전설 칭호 랭킹 기능 구현 - 내비게이션 타이틀 설정
